Update documentation of 'assemble' module for 'remote_src' and 'ignore_hidden'.
The assemble module documents the default value for 'remote_src' as 'True'.
Actually, it's implemented as 'False', so we correct this.

Ansible documentation commonly uses 'yes' and 'no' as the boolean choices,
so update the choices, description and default value to make this option
more consistent.

The 'ignore_hidden' option is also a boolean, but doesn't document its
choices, so we add them here.