Update choices and aliases in the apt module to be complete.
The apt module is missing documentation of the 'removed' and 'installed'
states (they are equivalent to the 'absent' and 'present' generic
terms used for state). These are added to the choices for that
option.

update_cache and install_recommends both have aliases where there is
a hypen in place of the underscore. These are added to the aliases
in the documentation.

The stderr and stdout options were specified as only returning in
the success case; actually these options can return in the failure
case.